Copy all field values when mapping InAppMessageEntity into InAppMessage models
CoreData objects should not be passed between threads. We ensure this by using the
private queue concurrency type on the managed object context, but we also map out
fields from the managed objects on to a data model.

These fields should be copied to ensure they remain available beyond the lifetime
of the NSManagedObject instance.

See: https://stackoverflow.com/q/13970807